Split up tests to speed up CI
