+++ test/Dataworker.loadData.fill.ts
@@ -37,7 +37,7 @@ import {
   bnZero,
 } from "../src/utils";
 import { MockConfigStoreClient, MockHubPoolClient, MockSpokePoolClient } from "./mocks";
-import { interfaces, utils as sdkUtils, constants as sdkConstants } from "@across-protocol/sdk";
+import { interfaces, constants as sdkConstants } from "@across-protocol/sdk";
 import { cloneDeep } from "lodash";
 import { CombinedRefunds } from "../src/dataworker/DataworkerUtils";
 import { INFINITE_FILL_DEADLINE } from "../src/common";
@@ -231,101 +231,6 @@ describe("Dataworker: Load data used in all functions", async function () {
       } as interfaces.FillWithBlock);
     }
 
-    it("Filters expired deposits", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send unexpired deposit
-      generateV3Deposit();
-      // Send expired deposit
-      const expiredDeposits = [generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 })];
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(2);
-      expect(
-        data1.expiredDepositsToRefundV3[originChainId][erc20_1.address].map((deposit) => deposit.depositId)
-      ).to.deep.equal(expiredDeposits.map((event) => event.args.depositId));
-      expect(data1.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(1);
-    });
-
-    it("Ignores disabled chains", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send unexpired deposit
-      generateV3Deposit();
-      // Send expired deposit
-      generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      // Returns no data if block range is undefined
-      const emptyData = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDisabledBlockRanges(),
-        spokePoolClients
-      );
-      expect(emptyData.bundleDepositsV3).to.deep.equal({});
-      expect(emptyData.expiredDepositsToRefundV3).to.deep.equal({});
-    });
-
-    it("Does not consider zero value deposits", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send unexpired and expired deposits in bundle block range.
-      generateV3Deposit({
-        inputAmount: bnZero,
-        message: "0x",
-      });
-      generateV3Deposit({
-        fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1,
-        inputAmount: bnZero,
-        message: "0x",
-      });
-
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-
-      expect(data1.bundleDepositsV3).to.deep.equal({});
-      expect(data1.expiredDepositsToRefundV3).to.deep.equal({});
-    });
-
-    it("Does not consider expired zero value deposits from prior bundle", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send expired deposits
-      const priorBundleDeposit = generateV3Deposit({
-        fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1,
-        inputAmount: bnZero,
-        message: "0x",
-      });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      const oldOriginChainToBlock = getDefaultBlockRange(5)[0][1];
-      const bundleBlockRanges = getDefaultBlockRange(5);
-      bundleBlockRanges[originChainIndex] = [priorBundleDeposit.blockNumber + 1, oldOriginChainToBlock];
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-
-      expect(data1.bundleDepositsV3).to.deep.equal({});
-      expect(data1.expiredDepositsToRefundV3).to.deep.equal({});
-    });
-
     it("Does not refund fills for zero value deposits", async function () {
       generateV3Deposit({
         inputAmount: bnZero,
@@ -393,86 +298,6 @@ describe("Dataworker: Load data used in all functions", async function () {
       );
 
       expect(data1.unexecutableSlowFills).to.deep.equal({});
-    });
-
-    it("Filters unexpired deposit out of block range", async function () {
-      // Send deposit behind and after origin chain block range. Should not be included in bundleDeposits.
-      // First generate mock deposit events with some block time between events.
-      const deposits = [
-        generateV3Deposit({ blockNumber: mockOriginSpokePoolClient.eventManager.blockNumber + 1 }),
-        generateV3Deposit({ blockNumber: mockOriginSpokePoolClient.eventManager.blockNumber + 11 }),
-        generateV3Deposit({ blockNumber: mockOriginSpokePoolClient.eventManager.blockNumber + 21 }),
-      ];
-      // Create a block range that contains only the middle deposit.
-      const originChainBlockRange = [deposits[1].blockNumber - 1, deposits[1].blockNumber + 1];
-      // Substitute origin chain bundle block range.
-      const bundleBlockRanges = [originChainBlockRange].concat(getDefaultBlockRange(5).slice(1));
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      expect(mockOriginSpokePoolClient.getDeposits().length).to.equal(deposits.length);
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(1);
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address][0].depositId).to.equal(deposits[1].args.depositId);
-    });
-    it("Includes duplicate deposits in bundle data", async function () {
-      generateV3Deposit();
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const depositToDuplicate = mockOriginSpokePoolClient.getDeposits()[0];
-      mockOriginSpokePoolClient.depositV3(depositToDuplicate);
-      mockOriginSpokePoolClient.depositV3(depositToDuplicate);
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      expect(
-        mockOriginSpokePoolClient.getDepositsForDestinationChainWithDuplicates(destinationChainId).length
-      ).to.equal(3);
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(3);
-    });
-    it("Filters duplicate deposits out of block range", async function () {
-      const deposit = generateV3Deposit({ blockNumber: mockOriginSpokePoolClient.eventManager.blockNumber + 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const depositToDuplicate = mockOriginSpokePoolClient.getDeposits()[0];
-      const duplicateDeposit = mockOriginSpokePoolClient.depositV3({
-        ...depositToDuplicate,
-        blockNumber: mockOriginSpokePoolClient.eventManager.blockNumber + 11,
-      });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      expect(
-        mockOriginSpokePoolClient.getDepositsForDestinationChainWithDuplicates(destinationChainId).length
-      ).to.equal(2);
-      const originChainBlockRange = [deposit.blockNumber, duplicateDeposit.blockNumber - 1];
-      // Substitute origin chain bundle block range.
-      const bundleBlockRanges = [originChainBlockRange].concat(getDefaultBlockRange(5).slice(1));
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(1);
-    });
-    it("Ignores expired deposits that were filled in same bundle", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send deposit that expires in this bundle.
-      const expiredDeposit = generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(1);
-      expect(data1.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(1);
-
-      // Now, send a fill for the deposit that would be in the same bundle. This should eliminate the expired
-      // deposit from a refund.
-      generateV3FillFromDepositEvent(expiredDeposit);
-      await mockDestinationSpokePoolClient.update(["FilledV3Relay"]);
-      const data2 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(6),
-        spokePoolClients
-      );
-      expect(data2.expiredDepositsToRefundV3).to.deep.equal({});
-      expect(data2.bundleFillsV3[repaymentChainId][l1Token_1.address].fills.length).to.equal(1);
     });
     it("Saves V3 fast fill under correct repayment chain and repayment token", async function () {
       const depositV3Events: Event[] = [];
@@ -854,141 +679,6 @@ describe("Dataworker: Load data used in all functions", async function () {
       expect(data1.bundleFillsV3[repaymentChainId][l1Token_1.address].fills.length).to.equal(1);
     });
 
-    it("Returns prior bundle expired deposits", async function () {
-      // Send deposit that expires in this bundle.
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      // Send deposit that expires in this bundle.
-      const expiredDeposit = generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(1);
-      expect(data1.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(1);
-
-      // Now, load a bundle that doesn't include the deposit in its range.
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      const oldOriginChainToBlock = getDefaultBlockRange(5)[0][1];
-      const bundleBlockRanges = getDefaultBlockRange(5);
-      bundleBlockRanges[originChainIndex] = [expiredDeposit.blockNumber + 1, oldOriginChainToBlock];
-      const data2 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-
-      // Now, there is no bundle deposit but still an expired deposit to refund.
-      expect(data2.bundleDepositsV3).to.deep.equal({});
-      expect(data2.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(1);
-    });
-    it("Includes duplicate prior bundle expired deposits", async function () {
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-
-      // Send deposit that expires in this bundle and send duplicate at same block.
-      const expiredDeposit = generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      mockOriginSpokePoolClient.depositV3(mockOriginSpokePoolClient.getDeposits()[0]);
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      expect(data1.bundleDepositsV3[originChainId][erc20_1.address].length).to.equal(2);
-      expect(data1.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(2);
-
-      // Now, load a bundle that doesn't include the deposits in its range.
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      const oldOriginChainToBlock = getDefaultBlockRange(5)[0][1];
-      const bundleBlockRanges = getDefaultBlockRange(5);
-      bundleBlockRanges[originChainIndex] = [expiredDeposit.blockNumber + 1, oldOriginChainToBlock];
-      const data2 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-
-      // Now, there is no bundle deposit but still two expired deposits to refund.
-      expect(data2.bundleDepositsV3).to.deep.equal({});
-      expect(data2.expiredDepositsToRefundV3[originChainId][erc20_1.address].length).to.equal(2);
-    });
-    it("Handles when deposit is greater than origin bundle end block but fill is within range", async function () {
-      // Send deposit after origin chain block range.
-      const blockRanges = getDefaultBlockRange(5);
-      const futureDeposit = generateV3Deposit();
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      blockRanges[originChainIndex] = [blockRanges[0][0], futureDeposit.blockNumber - 1];
-
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      generateV3FillFromDepositEvent(futureDeposit);
-      await mockDestinationSpokePoolClient.update(["FilledV3Relay"]);
-
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(blockRanges, spokePoolClients);
-      expect(data1.bundleDepositsV3).to.deep.equal({});
-
-      // Fill should not be included since we cannot validate fills when the deposit is in a following bundle.
-      // This fill is considered a "pre-fill" and will be validated when the deposit is included in a bundle.
-      expect(data1.bundleFillsV3).to.deep.equal({});
-      expect(spyLogIncludes(spy, -2, "invalid V3 fills in range")).to.be.true;
-    });
-    it("Does not count prior bundle expired deposits that were filled", async function () {
-      // Send deposit that expires in this bundle.
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      const expiredDeposit = generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-
-      // Let's make fill status for the relay hash always return Filled.
-      const expiredDepositHash = sdkUtils.getRelayHashFromEvent(mockOriginSpokePoolClient.getDeposits()[0]);
-      mockDestinationSpokePool.fillStatuses.whenCalledWith(expiredDepositHash).returns(interfaces.FillStatus.Filled);
-
-      // Now, load a bundle that doesn't include the deposit in its range.
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      const oldOriginChainToBlock = getDefaultBlockRange(5)[0][1];
-      const bundleBlockRanges = getDefaultBlockRange(5);
-      bundleBlockRanges[originChainIndex] = [expiredDeposit.blockNumber + 1, oldOriginChainToBlock];
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-
-      // There should be no expired deposit to refund because its fill status is Filled.
-      expect(data1.bundleDepositsV3).to.deep.equal({});
-      expect(data1.expiredDepositsToRefundV3).to.deep.equal({});
-    });
-    it("Does not count prior bundle expired deposits that we queried a fill for", async function () {
-      // Send deposit that expires in this bundle.
-      const bundleBlockTimestamps = await dataworkerInstance.clients.bundleDataClient.getBundleBlockTimestamps(
-        [originChainId, destinationChainId],
-        getDefaultBlockRange(5),
-        spokePoolClients
-      );
-      const expiredDeposit = generateV3Deposit({ fillDeadline: bundleBlockTimestamps[destinationChainId][1] - 1 });
-      await mockOriginSpokePoolClient.update(["V3FundsDeposited"]);
-      const deposits = mockOriginSpokePoolClient.getDeposits();
-
-      // Unlike previous test, we send a fill that the spoke pool client should query which also eliminates this
-      // expired deposit from being refunded.
-      generateV3FillFromDeposit(deposits[0]);
-      await mockDestinationSpokePoolClient.update(["RequestedV3SlowFill", "FilledV3Relay"]);
-      expect(mockDestinationSpokePoolClient.getFills().length).to.equal(1);
-
-      // Now, load a bundle that doesn't include the deposit in its range.
-      const originChainIndex = dataworkerInstance.chainIdListForBundleEvaluationBlockNumbers.indexOf(originChainId);
-      const oldOriginChainToBlock = getDefaultBlockRange(5)[0][1];
-      const bundleBlockRanges = getDefaultBlockRange(5);
-      bundleBlockRanges[originChainIndex] = [expiredDeposit.blockNumber + 1, oldOriginChainToBlock];
-      const data1 = await dataworkerInstance.clients.bundleDataClient.loadData(bundleBlockRanges, spokePoolClients);
-
-      // There should be no expired deposit to refund.
-      expect(data1.bundleDepositsV3).to.deep.equal({});
-      expect(data1.expiredDepositsToRefundV3).to.deep.equal({});
-      expect(data1.bundleFillsV3[repaymentChainId][l1Token_1.address].fills.length).to.equal(1);
-    });
-
     it("getBundleTimestampsFromCache and setBundleTimestampsInCache", async function () {
       // Unit test
       await dataworkerInstance.clients.bundleDataClient.loadData(getDefaultBlockRange(5), spokePoolClients);
@@ -1010,66 +700,6 @@ describe("Dataworker: Load data used in all functions", async function () {
   });
 
   describe("Miscellaneous functions", function () {
-    it("getUpcomingDepositAmount", async function () {
-      // Send two deposits on different chains
-      await depositV3(
-        spokePool_1,
-        destinationChainId,
-        depositor,
-        erc20_1.address,
-        amountToDeposit,
-        ZERO_ADDRESS,
-        amountToDeposit
-      );
-      await depositV3(
-        spokePool_2,
-        originChainId,
-        depositor,
-        erc20_2.address,
-        amountToDeposit,
-        ZERO_ADDRESS,
-        amountToDeposit
-      );
-      await updateAllClients();
-      expect(await bundleDataClient.getUpcomingDepositAmount(originChainId, erc20_1.address, 0)).to.equal(
-        amountToDeposit
-      );
-      expect(await bundleDataClient.getUpcomingDepositAmount(destinationChainId, erc20_2.address, 0)).to.equal(
-        amountToDeposit
-      );
-
-      // Removes deposits using block, token, and chain filters.
-      expect(
-        await bundleDataClient.getUpcomingDepositAmount(
-          originChainId,
-          erc20_1.address,
-          spokePoolClient_1.latestBlockSearched // block higher than the deposit
-        )
-      ).to.equal(0);
-      expect(
-        await bundleDataClient.getUpcomingDepositAmount(
-          originChainId,
-          erc20_2.address, // diff token
-          0
-        )
-      ).to.equal(0);
-      expect(
-        await bundleDataClient.getUpcomingDepositAmount(
-          destinationChainId, // diff chain
-          erc20_1.address,
-          0
-        )
-      ).to.equal(0);
-
-      // spoke pool client for chain not defined
-      expect(
-        await bundleDataClient.getUpcomingDepositAmount(
-          originChainId + destinationChainId + repaymentChainId + 1, // spoke pool client for chain is not defined in BundleDataClient
-          erc20_1.address,
-          0
-        )
-      ).to.equal(0);
-    });
     it("getApproximateRefundsForBlockRange", async function () {
       // Send two deposits on different chains
       // Fill both deposits and request repayment on same chain
